Refactor the file structure and naming of the elasticsearch-asset-apis
