+++ asset/src/elasticsearch_sender_api/api.ts
@@ -3,23 +3,34 @@ import {
     isNil, isString, isPlainObject, AnyObject, isNumber, getTypeOf
 } from '@terascope/utils';
 import elasticAPI from '@terascope/elasticsearch-api';
-import { createBulkSenderAPI, ElasticsearchSender } from '@terascope/elasticsearch-asset-apis';
+import {
+    createElasticsearchBulkSender,
+    ElasticsearchBulkSender
+} from '@terascope/elasticsearch-asset-apis';
 import { ElasticsearchAPISenderConfig } from './interfaces';
 
 export default class ElasticsearchSenderAPI extends APIFactory
-    <ElasticsearchSender, ElasticsearchAPISenderConfig> {
+    <ElasticsearchBulkSender, ElasticsearchAPISenderConfig> {
     // TODO: there might need more checks here
     validateConfig(config: unknown): ElasticsearchAPISenderConfig {
-        if (isNil(config)) throw new Error('No configuration was found or provided for elasticsearch_reader_api');
-        if (!isObject(config)) throw new Error(`Invalid config, must be an object, was given ${getTypeOf(config)}`);
-        if (!isNumber(config.size)) throw new Error(`Invalid size parameter, expected number, got ${getTypeOf(config.size)}`);
-        if (isNil(config.connection) || !isString(config.connection)) throw new Error('Invalid parameter "connection", must provide a valid connection');
+        if (isNil(config)) {
+            throw new Error('No configuration was found or provided for elasticsearch_reader_api');
+        }
+        if (!isObject(config)) {
+            throw new Error(`Invalid config, must be an object, was given ${getTypeOf(config)}`);
+        }
+        if (!isNumber(config.size)) {
+            throw new Error(`Invalid size parameter, expected number, got ${getTypeOf(config.size)}`);
+        }
+        if (isNil(config.connection) || !isString(config.connection)) {
+            throw new Error('Invalid parameter "connection", must provide a valid connection');
+        }
         return config as ElasticsearchAPISenderConfig;
     }
 
     async create(
         _name: string, overrideConfig: Partial<ElasticsearchAPISenderConfig>
-    ): Promise<{ client: ElasticsearchSender, config: ElasticsearchAPISenderConfig }> {
+    ): Promise<{ client: ElasticsearchBulkSender, config: ElasticsearchAPISenderConfig }> {
         const apiConfig = this.validateConfig(Object.assign({}, this.apiConfig, overrideConfig));
         const { api_name, ...config } = apiConfig;
 
@@ -30,7 +41,7 @@ export default class ElasticsearchSenderAPI extends APIFactory
         });
 
         const esClient = elasticAPI(client, this.context.logger, config);
-        const sender = createBulkSenderAPI({ client: esClient, config });
+        const sender = createElasticsearchBulkSender({ client: esClient, config });
 
         return { client: sender, config: apiConfig };
     }
