+++ packages/elasticsearch-asset-apis/src/elasticsearch-bulk-sender/ElasticsearchBulkSender.ts
@@ -9,9 +9,9 @@ import {
 } from '@terascope/utils';
 import {
     ElasticsearchSenderConfig, IndexSpec, BulkMeta, UpdateConfig
-} from '../interfaces';
+} from './interfaces';
 
-export class ElasticsearchSender implements RouteSenderAPI {
+export class ElasticsearchBulkSender implements RouteSenderAPI {
     client: elasticAPI.Client;
     config: ElasticsearchSenderConfig;
     clientVersion: number;
