+++ packages/elasticsearch-asset-apis/src/elasticsearch-bulk-sender/interfaces.ts
@@ -1,0 +1,44 @@
+import { AnyObject } from '@terascope/types';
+
+export interface ElasticsearchSenderConfig {
+    size: number;
+    connection: string;
+    index: string;
+    type?: string;
+    delete?: boolean;
+    update?: boolean;
+    update_retry_on_conflict?: number;
+    update_fields?: string[];
+    upsert?: boolean;
+    create?: boolean;
+    script_file?: string;
+    script?: string;
+    script_params?: AnyObject;
+    _key?: string
+}
+
+export interface BulkMeta {
+    _index: string;
+    _type: string;
+    _id: string | number;
+    retry_on_conflict: number;
+}
+
+export interface IndexSpec {
+    index?: AnyObject;
+    create?: AnyObject;
+    update?: AnyObject;
+    delete?: AnyObject;
+}
+
+export interface ScriptConfig {
+    file?: string;
+    source?: string;
+    params?: AnyObject;
+}
+
+export interface UpdateConfig {
+    upsert?: AnyObject;
+    doc?: AnyObject;
+    script?: ScriptConfig;
+}
