+++ packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/algorithms/date-helpers.ts
