+++ packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/algorithms/id-slicer.ts
