+++ packages/elasticsearch-asset-apis/src/elasticsearch-reader-api/interfaces.ts
@@ -1,8 +1,79 @@
+import type { DataFrame } from '@terascope/data-mate';
+import type { AnyObject, DataEntity, Logger } from '@terascope/utils';
+import type {
+    SearchParams
+} from 'elasticsearch';
 import type { LifeCycle, SlicerRecoveryData } from '@terascope/job-components';
-import { EventEmitter } from 'events';
-import { AnyObject, Logger } from '@terascope/utils';
-import { DataTypeConfig, xLuceneVariables } from '@terascope/types';
-import { WindowState } from './window-state';
+import type { EventEmitter } from 'events';
+import type { DataTypeConfig, xLuceneVariables } from '@terascope/types';
+import type { WindowState } from './WindowState';
+
+/**
+ * This is used for as the internal elasticsearch
+ * client in reader, this is designed as an abstraction
+ * so that spaces client will work with it is own specific
+ * optimizations
+*/
+export interface ReaderClient {
+    /**
+     * Counts the number of documents for a given query
+    */
+    count(query: SearchParams): Promise<number>;
+
+    /**
+     * Searches for documents for a given query
+    */
+    search(
+        query: SearchParams,
+        useDataFrames: false,
+        typeConfig?: DataTypeConfig
+    ): Promise<DataEntity[]>;
+    search(
+        query: SearchParams,
+        useDataFrames: true,
+        typeConfig: DataTypeConfig
+    ): Promise<DataFrame>;
+    search(
+        query: SearchParams,
+        useDataFrames: boolean,
+        typeConfig?: DataTypeConfig
+    ): Promise<DataEntity[]|DataFrame>;
+
+    /**
+     * Used to make a custom search request
+     *
+     * @note this API is subject to change
+    */
+    _searchRequest(query: SearchParams, fullResponse?: false): Promise<DataEntity[]>;
+    _searchRequest(query: SearchParams, fullResponse: true): Promise<unknown>;
+    _searchRequest(query: SearchParams, fullResponse?: boolean): Promise<DataEntity[]|unknown>;
+
+    /**
+     * Gets the elasticsearch major server version,
+     * this will be used to format the search parameters
+    */
+    getESVersion(): number;
+
+    /**
+     * Verify that the cluster is up,
+     * internally this will use esClient.version() probably
+    */
+    verify(): Promise<void>;
+
+    /**
+     * Used to determine the max window size
+    */
+    getSettings(index: string): Promise<SettingResults>;
+}
+
+export interface SettingResults {
+    [key: string]: {
+        settings: {
+            'index.max_result_window': number
+        },
+        defaults: AnyObject
+    }
+}
 
 export enum IDType {
     base64url = 'base64url',
@@ -188,47 +259,4 @@ export interface DetermineSliceResults {
 export interface SlicerDateConfig extends DateSegments {
     end: moment.Moment;
     holes?: DateConfig[];
-}
-
-export interface ElasticsearchSenderConfig {
-    size: number;
-    connection: string;
-    index: string;
-    type?: string;
-    delete?: boolean;
-    update?: boolean;
-    update_retry_on_conflict?: number;
-    update_fields?: string[];
-    upsert?: boolean;
-    create?: boolean;
-    script_file?: string;
-    script?: string;
-    script_params?: AnyObject;
-    _key?: string
-}
-
-export interface BulkMeta {
-    _index: string;
-    _type: string;
-    _id: string | number;
-    retry_on_conflict: number;
-}
-
-export interface IndexSpec {
-    index?: AnyObject;
-    create?: AnyObject;
-    update?: AnyObject;
-    delete?: AnyObject;
-}
-
-export interface ScriptConfig {
-    file?: string;
-    source?: string;
-    params?: AnyObject;
-}
-
-export interface UpdateConfig {
-    upsert?: AnyObject;
-    doc?: AnyObject;
-    script?: ScriptConfig;
 }
