+++ packages/elasticsearch-asset-apis/src/index.ts
@@ -1,11 +1,2 @@
 export * from './elasticsearch-reader-api';
-export * from './elasticsearch-api';
 export * from './elasticsearch-bulk-sender';
-export * from './elasticsearch-date-slicer/helpers';
-export * from './elasticsearch-id-slicer/helpers';
-export * from './elasticsearch-bulk-sender/bulk-sender';
-export * from './ElasticsearchReaderClient';
-export * from './reader-client';
-export * from './interfaces';
-export * from './spaces-api';
-export * from './window-state';
